Checkpointing
Found an inconsistency in the encoding of the nested struct array between the JS and solidity.
Seems solidity is encoding an empty array as a lot of zeroes. For some reason JS is encoding it as something else.

Example coding of a delegation message in JS:
> 0x409f5114779a253e700d775d7845e6efc1e83685ac59868d2df3d4de51c7d62100000000000000000000000070997970c51812dc3a010c7d01b50e0d17dc79c80000000000000000000000000000000000000000000000000000000000000000c5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470

Delegation encoding in solidity:
> 0x409f5114779a253e700d775d7845e6efc1e83685ac59868d2df3d4de51c7d62100000000000000000000000070997970c51812dc3a010c7d01b50e0d17dc79c8000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000000